new arguments storage structure
