+++ gateway/core/services/runners/fleets_runner.py
@@ -35,7 +35,7 @@ from core.ibm_cloud.code_engine.fleets.utils import (
     build_run_env_variables,
     build_run_volume_mounts,
 )
-from core.services.storage.arguments_storage import ArgumentsStorage
+from core.services.storage import get_arguments_storage
 
 logger = logging.getLogger("FleetsRunner")
 
@@ -575,7 +575,7 @@ class FleetsRunner(AbstractRunner):
             paths: Dict from :meth:`_build_cos_paths`.
         """
         program = self.job.program
-        storage = ArgumentsStorage(self.job.author.username, program)
+        storage = get_arguments_storage(self.job.author.username, program)
         content = storage.get(str(self.job.id)) or "{}"
 
         try:
